Move complex schema holder class into own module
